Move TrailingZeroCountFallback to common SpanHelpers
So it can be used by other types than byte
