+++ src/System.Private.CoreLib/shared/System/SpanHelpers.Byte.cs
@@ -1244,18 +1244,5 @@ namespace System
                                                        0x03ul << 32 |
                                                        0x02ul << 40 |
                                                        0x01ul << 48) + 1;
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int TrailingZeroCountFallback(int matches)
-        {
-            // https://graphics.stanford.edu/~seander/bithacks.html#ZerosOnRightMultLookup
-            return TrailingCountMultiplyDeBruijn[(int)(((uint)((matches & -matches) * 0x077CB531U)) >> 27)];
-        }
-
-        private static ReadOnlySpan<byte> TrailingCountMultiplyDeBruijn => new byte[32]
-        {
-            0, 1, 28, 2, 29, 14, 24, 3, 30, 22, 20, 15, 25, 17, 4, 8,
-            31, 27, 13, 23, 21, 19, 16, 7, 26, 12, 18, 6, 11, 5, 10, 9
-        };
     }
 }
